pyproject.toml
[project]
name = "opensearch-keeper"
version = "0.1.0"
description = "A CLI tool for managing OpenSearch index templates"
keywords = ["opensearch", "elasticsearch", "templates", "cli"]
readme = "README.md"
authors = [
    {name = "krushik", email = "krushik@gmail.com"},
]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
requires-python = ">=3.7"
dependencies = [
    "PyYAML>=6.0",
    "boto3>=1.24.0",
    "click>=8.0.0",
    "opensearch-py>=2.0.0",
    "requests[socks]>=2.27.0",
]

[project.urls]
Homepage = "https://github.com/krushik/opensearch-keeper"

[project.scripts]
opensearch-keeper = "opensearch_keeper.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.uv]
dev-dependencies = [
    "boto3-stubs[opensearch]>=1.34.4",
    "mypy>=1.4.1",
    "mypy-boto3-opensearch>=1.34.1",
    "ruff>=0.11.2",
    "types-pyyaml>=6.0.12.12",
    "types-requests>=2.31.0.6",
    "pytest>=7.4.4",
    "pip>=24.0",
]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
